modules/sst_gen/stt.py
import os
from dotenv import load_dotenv

load_dotenv()
api_key = os.getenv("OPENAI_API_KEY")

from openai import OpenAI
client = OpenAI()

#!  add local vs api/web based switch

def transcribe_GPT(audio_file_path):
    with open(audio_file_path, "rb") as audio_file:
        transcript = client.audio.transcriptions.create(
            model="whisper-1",
            file=audio_file
        )
        return transcript.text

if __name__ == "__main__":
    print(transcribe_GPT("path/to/audio/file.mp3"))