requirements.txt
openai
langchain
langchain_openai
langchain_experimental
langchain_core
langsmith
langchain_text_splitters
exa_py
duckduckgo-search
python-dotenv
notebook
tiktoken
lxml
pypdf
faiss-cpu
Pillow
markdown
customtkinter
streamlit
ollama
networkx 
matplotlib 
numpy
transformers
